# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Baika Service",
        "version": "1.0.0"
    },
    "paths": {
        "/api/automobiles": {
            "get": {
                "tags": [
                    "Api Baika Service"
                ],
                "summary": "Route automobiles",
                "description": "Retourne toutes les automobiles",
                "operationId": "automobiles",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Vous n'êtes pas autorisé à consulter cette page."
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "La ressource demandée est introuvable."
                    }
                }
            }
        },
        "/api/automobile/{automobile_id}": {
            "get": {
                "tags": [
                    "Api Baika Service"
                ],
                "summary": "Route automobiles Image",
                "description": "Retourne une automobile",
                "operationId": "automobile Single",
                "parameters": [
                    {
                        "name": "automobile_id",
                        "in": "path",
                        "description": "ID of User",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "e1c26f26-7c78-4d06-85c0-132052644648"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Vous n'êtes pas autorisé à consulter cette page."
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "La ressource demandée est introuvable."
                    }
                }
            }
        },
        "/api/location/{location_id}/images": {
            "get": {
                "tags": [
                    "Api Baika Service"
                ],
                "summary": "Route locations Image",
                "description": "Retourne toutes les limages d'une location",
                "operationId": "locationsImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Pour la obtenir les images",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Vous n'êtes pas autorisé à consulter cette page."
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "La ressource demandée est introuvable."
                    }
                }
            }
        },
        "/api/location/{location_id}": {
            "get": {
                "tags": [
                    "Api Baika Service"
                ],
                "summary": "Route locations Image",
                "description": "Retourne une location",
                "operationId": "location Single",
                "parameters": [
                    {
                        "name": "location_id",
                        "in": "path",
                        "description": "ID of User",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "e1c26f26-7c78-4d06-85c0-132052644648"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Vous n'êtes pas autorisé à consulter cette page."
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "La ressource demandée est introuvable."
                    }
                }
            }
        },
        "/api/locations": {
            "get": {
                "tags": [
                    "Api Baika Service"
                ],
                "summary": "Route locations",
                "description": "Retourne toutes les locations",
                "operationId": "locations",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Vous n'êtes pas autorisé à consulter cette page."
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "La ressource demandée est introuvable."
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}